requirements.txt
# AI Recruitment Assistant - Python Dependencies
# Optimized for RTX 4060 (8GB VRAM) and Windows

# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Hugging Face Ecosystem
transformers>=4.35.0
datasets>=2.14.0
tokenizers>=0.14.0

# Parameter-Efficient Fine-Tuning
peft>=0.6.0

# Training Acceleration
accelerate>=0.24.0

# Quantization (4-bit for memory efficiency)
bitsandbytes>=0.41.0

# Experiment Tracking & Monitoring
wandb>=0.16.0
tensorboard>=2.14.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
scikit-learn>=1.3.0

# Configuration Management
pyyaml>=6.0
omegaconf>=2.3.0

# Development Tools (Optional)
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API and Web Framework (for inference server)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Additional ML Utilities
scipy>=1.11.0
requests>=2.31.0
python-multipart>=0.0.6

# Note: For CUDA support, run after pip install:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
